Add Projects.md and semester 1 projects and boards
The focus of the learning time is built around projects. We asked
candidates to create new github repositories for each project
individually, and also create a project board on github.

Github Projects aren't usually used at GDS, however in this case as we
have candidates spread across GDS teams that tend to use different tools
(Trello / Jira etc), Github Projects was the easiest tool that we all
already had access to.

This adds project boards and links to repos for semester 1 candidates. A
couple are missing who will join or add data later, I will update this
as we go.