Rearrange folder structure for presentations
I think it's best to make a folder by date so we end up with:

- 2023-01-01
- 2023-01-02
- etc..

and then each contains the PDFs for a day grouped together with a
sortable title.

I've left unfinished markdown files in here, PDFs are not accessibly
friendly, but right now I can't think of anything else to put in version
control. We should link back to the google slides for internal GDS folks
with links, and can put speaker notes / transcripts into the markdown
files if we get around to it.